[Refac] FPlayerStat의 HP,Atk_Damage 자료형을 int->float으로 수정했습니다.

+++ TeamName/Player/Player.h
@@ -8,9 +8,9 @@ class Inventory;
 class Monster;
 
 struct FPlayerStat{
-    unsigned int HP = 200; // PlayerHP
+    float HP = 200.f; // PlayerHP
     unsigned int Stamina = 0; // PlayerStamina
-	unsigned int Atk_Damage = 30; // Player Attack Damage
+	float Atk_Damage = 30; // Player Attack Damage
 	unsigned int Level = 1; // PlayerLevel
 	unsigned short EXP = 0; // PlayerEXP
     std::string name = "None"; // PlayerName
@@ -18,7 +18,7 @@ struct FPlayerStat{
 };
 class IPlayer{
 public:
-    virtual void InitializeStat(int _hp, std::string _name, int _atkdamage,int _level,int _stamina = 0) = 0;
+    virtual void InitializeStat(float _hp, std::string _name, float _atkdamage,int _level,int _stamina = 0) = 0;
     virtual const FPlayerStat& Getstat() = 0;
     virtual FPlayerStat& SetStat() = 0;
     virtual void Attack(Monster* _monster) = 0;
@@ -33,7 +33,7 @@ public:
 	// 언리얼엔진의 델리게이트랑 비슷하다 생각하시면 됩니다.
 	std::function<void()> OnDead;
     /*****Stat*****/
-    void InitializeStat(int _hp,std::string _name,int _atkdamage, int _level,int _stamina = 0) override;
+    void InitializeStat(float _hp,std::string _name,float _atkdamage, int _level,int _stamina = 0) override;
     _forceinline const FPlayerStat& Getstat() override{return Playerstat;}
     _forceinline FPlayerStat& SetStat() override {return Playerstat;}
 	void AcquireEXP(int _exp);
